+++ src/components/App/CardGrid/index.tsx
@@ -1,9 +1,9 @@
 import * as React from 'react';
 
 import { Card } from '../../Global';
-import type { CardProps } from '../../Global';
 
 import styles from './CardGrid.module.scss';
+import { CardProps } from '../../../types/cards';
 
 const CardGrid = ({ cardList }: { cardList: CardProps[] }) => {
   return (
