+++ src/components/Global/Logo/index.tsx
@@ -2,11 +2,7 @@ import React from 'react';
 import Image from 'next/image';
 import logo from '../../../../public/assets/logo.jpg';
 import logoStyles from './Logo.module.scss';
-
-type LogoProps = {
-  showMark: boolean;
-  showType: boolean;
-};
+import { LogoProps } from '../../../types/global';
 
 const Logo = ({ showMark, showType }: LogoProps) => {
   return (
