+++ src/services/signup.ts
@@ -1,5 +1,5 @@
 import Axios from './axios';
-import type { signUpData } from '../pages/api/auth/signup';
+import type { signUpData } from '../types/api/auth/sign-up';
 
 /**
  * This will send an axios request to the server to sign up a user
