+++ src/types/api/auth/reset-password.ts
@@ -1,0 +1,5 @@
+export type resetPasswordProps = {
+  email: string;
+  oldPassword: string;
+  newPassword: string;
+};
