+++ src/types/api/auth/sign-up.ts
@@ -1,0 +1,4 @@
+export type signUpData = {
+  email: string;
+  password: string;
+};
